Refactor: merge HasStorage back into Node
