+++ pyiron_workflow/node.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 from abc import ABC
 from concurrent.futures import Future
+from importlib import import_module
 from typing import Any, Literal, Optional, TYPE_CHECKING
 
 from pyiron_snippets.colors import SeabornColors
@@ -21,7 +22,7 @@ from pyiron_workflow.mixin.injection import HasIOWithInjection
 from pyiron_workflow.mixin.run import Runnable, ReadinessError
 from pyiron_workflow.mixin.semantics import Semantic
 from pyiron_workflow.mixin.single_output import ExploitsSingleOutput
-from pyiron_workflow.mixin.storage import HasPickleStorage
+from pyiron_workflow.mixin.storage import StorageInterface, PickleStorage
 from pyiron_workflow.topology import (
     get_nodes_in_data_tree,
     set_run_connections_according_to_linear_dag,
@@ -45,7 +46,6 @@ class Node(
     HasIOWithInjection,
     ExploitsSingleOutput,
     HasWorkingDirectory,
-    HasPickleStorage,
     ABC,
 ):
     """
@@ -295,8 +295,9 @@ class Node(
         super().__init__(
             label=self.__class__.__name__ if label is None else label,
             parent=parent,
-            storage_backend=storage_backend,
         )
+        self._storage_backend = None
+        self.storage_backend = storage_backend
         self.save_after_run = save_after_run
         self.cached_inputs = None
         self._user_data = {}  # A place for power-users to bypass node-injection
@@ -781,6 +782,18 @@ class Node(
         else:
             logger.info(f"Could not replace_child {self.label}, as it has no parent.")
 
+    @classmethod
+    def _storage_interfaces(cls):
+        return {"pickle": PickleStorage}
+
+    @classmethod
+    def allowed_backends(cls):
+        return tuple(cls._storage_interfaces().keys())
+
+    @classmethod
+    def default_backend(cls):
+        return "pickle"
+
     @property
     def storage_directory(self) -> DirectoryObject:
         return self.working_directory
@@ -791,3 +804,144 @@ class Node(
 
     def tidy_storage_directory(self):
         self.tidy_working_directory()
+
+    _save_load_warnings = """
+        HERE BE DRAGONS!!!
+
+        Warning:
+            This almost certainly only fails for subclasses of :class:`Node` that don't
+            override `node_function` or `macro_creator` directly, as these are expected 
+            to be part of the class itself (and thus already present on our instantiated 
+            object) and are never stored. Nodes created using the provided decorators 
+            should all work.
+
+        Warning:
+            If you modify a `Macro` class in any way (changing its IO maps, rewiring 
+            internal connections, or replacing internal nodes), don't expect 
+            saving/loading to work.
+
+        Warning:
+            If the underlying source code has changed since saving (i.e. the node doing 
+            the loading does not use the same code as the node doing the saving, or the 
+            nodes in some node package have been modified), then all bets are off.
+
+    """
+
+    def save(self):
+        """
+        Writes the node to file (using HDF5) such that a new node instance of the same
+        type can :meth:`load()` the data to return to the same state as the save point,
+        i.e. the same data IO channel values, the same flags, etc.
+        """
+        self.storage.save(self)
+
+    save.__doc__ += _save_load_warnings
+
+    def load(self):
+        """
+        Loads the node file (from HDF5) such that this node restores its state at time
+        of loading.
+
+        Raises:
+            TypeError: when the saved node has a different class name.
+        """
+        if self.storage.has_contents(self):
+            self.storage.load(self)
+        else:
+            # Check for saved content using any other backend
+            for backend in self.allowed_backends():
+                interface = self._storage_interfaces()[backend]()
+                if interface.has_contents(self):
+                    interface.load(self)
+                    break
+
+    load.__doc__ += _save_load_warnings
+
+    def delete_storage(self):
+        """Remove save files for _all_ available backends."""
+        for backend in self.allowed_backends():
+            interface = self._storage_interfaces()[backend]()
+            try:
+                interface.delete(self)
+            except FileNotFoundError:
+                pass
+
+    @property
+    def storage_root(self):
+        """The parent-most object that has storage."""
+        parent = self.parent
+        if isinstance(parent, Node):
+            return parent.storage_root
+        else:
+            return self
+
+    @property
+    def storage_backend(self):
+        storage_root = self.storage_root
+        if storage_root is self:
+            backend = self._storage_backend
+        else:
+            backend = storage_root.storage_backend
+        return self.default_backend() if backend is None else backend
+
+    @storage_backend.setter
+    def storage_backend(self, new_backend):
+        storage_root = self.storage_root
+        if new_backend is not None:
+            if new_backend not in self.allowed_backends():
+                raise ValueError(
+                    f"{self.label} got the storage backend {new_backend}, but only "
+                    f"{self.allowed_backends()} are permitted."
+                )
+            elif (
+                storage_root is not self and new_backend != storage_root.storage_backend
+            ):
+                raise ValueError(
+                    f"Storage backends should only be set on the storage root "
+                    f"({self.storage_root.label}), not on child ({self.label})"
+                )
+        self._storage_backend = new_backend
+
+    @property
+    def storage(self) -> StorageInterface:
+        if self.storage_backend is None:
+            raise ValueError(f"{self.label} does not have a storage backend set")
+        return self._storage_interfaces()[self.storage_backend]()
+
+    @property
+    def any_storage_has_contents(self):
+        return any(
+            self._storage_interfaces()[backend]().has_contents(self)
+            for backend in self.allowed_backends()
+        )
+
+    @property
+    def import_ready(self) -> bool:
+        """
+        Checks whether `importlib` can find this node's class, and if so whether the
+        imported object matches the node's type.
+
+        Returns:
+            (bool): Whether the imported module and name of this node's class match
+                its type.
+        """
+        try:
+            module = self.__class__.__module__
+            class_ = getattr(import_module(module), self.__class__.__name__)
+            if module == "__main__":
+                logger.warning(f"{self.label} is only defined in __main__")
+            return type(self) is class_
+        except (ModuleNotFoundError, AttributeError):
+            return False
+
+    @property
+    def import_readiness_report(self):
+        print(self.report_import_readiness())
+
+    def report_import_readiness(self, tabs=0, report_so_far=""):
+        newline = "\n" if len(report_so_far) > 0 else ""
+        tabspace = tabs * "\t"
+        return (
+            report_so_far + f"{newline}{tabspace}{self.label}: "
+                            f"{'ok' if self.import_ready else 'NOT IMPORTABLE'}"
+        )
